// CodeGenerator/models/pagination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 서울 따릉이 시뮬레이션 플랫폼 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Pagination
 */
export interface Pagination {
    /**
     * 현재 페이지 번호
     * @type {number}
     * @memberof Pagination
     */
    'currentPage': number;
    /**
     * 전체 페이지 수
     * @type {number}
     * @memberof Pagination
     */
    'totalPages': number;
    /**
     * 전체 항목 수
     * @type {number}
     * @memberof Pagination
     */
    'totalItems': number;
    /**
     * 페이지당 항목 수
     * @type {number}
     * @memberof Pagination
     */
    'itemsPerPage': number;
    /**
     * 다음 페이지 존재 여부
     * @type {boolean}
     * @memberof Pagination
     */
    'hasNext': boolean;
    /**
     * 이전 페이지 존재 여부
     * @type {boolean}
     * @memberof Pagination
     */
    'hasPrev': boolean;
}